Fix patch to install TinyTeX relative to conda environment
Previously, the entire conda environment was being overwritten by the
command `quarto install tinytex`.

This required an additional hack in the code for detecting TinyTeX:

When `hasTexLive` was called, it was detecting an installation, but
`quarto` was still looking for `tlmgr` at the patched TinyTeX install
location. So I simplified that function to only look for the
Quarto-installed version of TinyTeX.